+++ test/common/services/saved_object_info/saved_object_info.ts
@@ -12,8 +12,9 @@ import { Either, fromNullable, chain, getOrElse, toError } from 'fp-ts/Either';
 import { flow, pipe } from 'fp-ts/function';
 import * as TE from 'fp-ts/lib/TaskEither';
 import * as T from 'fp-ts/lib/Task';
+import { inspect } from 'util';
+import { ToolingLog } from '@kbn/dev-utils';
 import { FtrService } from '../../ftr_provider_context';
-import { format } from './utils';
 
 const pluck = (key: string) => (obj: any): Either<Error, string> =>
   fromNullable(new Error(`Missing ${key}`))(obj[key]);
@@ -55,11 +56,17 @@ export const types = (node: string) => async (index: string = '.kibana') =>
 export class SavedObjectInfoService extends FtrService {
   private readonly config = this.ctx.getService('config');
 
-  public getTypes() {
-    return types(url.format(this.config.get('servers.elasticsearch')));
-  }
+  public async logSoTypes(log: ToolingLog, msg: string | null = null) {
+    const print = (xs: any) =>
+      log.info(
+        `\n### Saved Object Types ${msg || 'Count: ' + xs.length}\n${inspect(xs, {
+          compact: false,
+          depth: 99,
+          breakLength: 80,
+          sorted: true,
+        })}`
+      );
 
-  public async getTypesPretty() {
-    return format(await this.getTypes()());
+    pipe(await types(url.format(this.config.get('servers.elasticsearch'))), print);
   }
 }
